perf(FlatDB): variable-width B-tree value slots + ulong pointers
B-tree index nodes now store per-entry value pointers (leaf MetadataStart,
intermediate ChildOffset) as variable 1..8 byte LE unsigned integers, with
slot width chosen per node to fit (max - BaseOffset). BaseOffset becomes a
mandatory fixed 6-byte LE field (256 TiB cap), and ValueSize narrows from
u16 to u8 since per-entry slots are at most 8 bytes. Hash-index variant
stays at 4-byte slots (writer guards MetadataStart > 4 GiB).

Also drops the unused 3-byte Uniform-key promotion in the layout planner
and a stale "BTree/Inline" comment in HsstEnumerator.

+++ src/Nethermind/Nethermind.State.Flat/Hsst/HsstIndex.cs
@@ -30,7 +30,7 @@ public readonly ref struct HsstIndex
 
     public ReadOnlySpan<byte> GetKey(int index) => _inner.GetKey(index);
     public ReadOnlySpan<byte> GetValue(int index) => _inner.GetValue(index);
-    public int GetIntValue(int index) => _inner.GetIntValue(index);
+    public ulong GetUInt64Value(int index) => _inner.GetUInt64Value(index);
     public int FindFloorIndex(ReadOnlySpan<byte> key) => _inner.FindFloorIndex(key);
     public int GetFullKey(int index, Span<byte> dest) => _inner.GetFullKey(index, dest);
